fix: clippy unnecessary_cast, profile i18n, admin style consistency, and admin API features
- Fix 8 unnecessary_cast warnings where row.get::<i64> now returns i64 directly
- Wire all remaining hardcoded strings in profile_page.dart to l10n
- Rewrite admin_page.dart with AppTheme consistency (colors, spacing, radius)
- Add admin order status update endpoint (POST /api/admin/orders/:id/status)
- Add admin user role change endpoint (POST /api/admin/users/:id/role)
- Add Flutter API methods: updateAdminOrderStatus, updateUserRole
- Update admin_stats to include category breakdown

+++ src/api/negotiate.rs
@@ -72,7 +72,7 @@ pub async fn list_negotiations(
             listing_id: row.get("listing_id"),
             buyer_id: row.get("buyer_id"),
             seller_id: row.get("seller_id"),
-            proposed_price: crate::utils::cents_to_yuan(row.get::<i32, _>("proposed_price") as i64),
+            proposed_price: crate::utils::cents_to_yuan(row.get::<i64, _>("proposed_price")),
             reason: row.get("reason"),
             status: row.get("status"),
             counter_price: row
@@ -174,7 +174,7 @@ pub async fn respond_negotiation(
         .fetch_one(&state.db)
         .await
         .map_err(|e| ApiError::Internal(anyhow::anyhow!("DB error: {}", e)))?;
-    let proposed_price: i64 = hitl_row.get::<i32, _>("proposed_price") as i64;
+    let proposed_price: i64 = hitl_row.get::<i64, _>("proposed_price") as i64;
 
     // Notify the buyer about the seller's decision
     let (notif_title, notif_body): (String, String) = match new_status {
